Rename proto_values->values, create values.test.ts (#2750)
